New additions
Added the flat list import and then made new style items for the box and the buttons.
Deleted the styles  that are customized in the  tags added styles from the style sheet instead which made. Then made stored the dishes item in a the flat list tag so it show every thing.