New Additions
Made the course item into a string .
Then added a new function which allow the course to update the array with courses name and the i re organized my array for me to  see better.
Then updated the array and the useState components to have the course as  enw component for the addition of the items in the array. Made it so  that the price is entered as a number specifically fraction/float.  Then added another scroll View to the top of the document.  First i installed react native paper for the use of he menu item and the Paper provider, them used the item.  Changed the margin of the box component of the the array view so that the items are separate, and clearly visible. Then Added the radius and the font weight of the delete button, but will still fix that. Adde some new styles for the menu and the add dish button,  also changed the font of certain components to Georgia since it is the easiest available font for me to use.