New  updates
Adde new style for the drop down menu.
Added a view for the entry point of information to edit the array, i.e., the name input, the price input, description and the drop down menu. The view covers all the afore mentioned components which are mentions and will put them in separate box. them Change  the dimensions of the box and and other entry point in the style sheet.  to create box for the view. Changed the name of restaurant into a Christoffelson's Barne.